+++ src/utils/constants.rs
@@ -5,6 +5,12 @@
 use crate::repositories::metadata::WEB_METADATA_PROVIDER_ID;
 
 pub const REGISTER_FILENAME: &str = "Register.toml";
+
+/// Note that when the `integration-tests` feature flag is enabled `CONFIG_FILENAME` is compiled differently.
+#[cfg(feature = "integration-tests")]
+pub const CONFIG_FILENAME: &str = "TestConfig.toml";
+
+#[cfg(not(feature = "integration-tests"))]
 pub const CONFIG_FILENAME: &str = "Config.toml";
 
 pub const DEFAULT_METADATA_REPOSITORY_NAME: &str = "core";