+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased](https://github.com/pack-it/packit/compare/0.0.4...HEAD)
 
+### Added
+- The integration tests to test Packit.
+
 ### Fixed
 - Fix patch apply directory meta check reporting wrong issues.